+++ src/test/kotlin/devtests/Or.kt
@@ -1,0 +1,11 @@
+package devtests
+
+import klib.extensions.or
+import klib.extensions.orNullable
+
+fun main() {
+    val test: String? = null
+
+    println(test or "not null")
+    println(test orNullable  null)
+}